+++ app/src/main/java/com/example/user/weatherapp/pojo/city_pojo/SysCity.java
@@ -8,10 +8,10 @@ public class SysCity {
 
     @SerializedName("type")
     @Expose
-    private Integer type;
+    private Double type;
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private Double id;
     @SerializedName("message")
     @Expose
     private Double message;
@@ -20,24 +20,24 @@ public class SysCity {
     private String country;
     @SerializedName("sunrise")
     @Expose
-    private Integer sunrise;
+    private Double sunrise;
     @SerializedName("sunset")
     @Expose
-    private Integer sunset;
+    private Double sunset;
 
-    public Integer getType() {
+    public Double getType() {
         return type;
     }
 
-    public void setType(Integer type) {
+    public void setType(Double type) {
         this.type = type;
     }
 
-    public Integer getId() {
+    public Double getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Double id) {
         this.id = id;
     }
 
@@ -57,19 +57,19 @@ public class SysCity {
         this.country = country;
     }
 
-    public Integer getSunrise() {
+    public Double getSunrise() {
         return sunrise;
     }
 
-    public void setSunrise(Integer sunrise) {
+    public void setSunrise(Double sunrise) {
         this.sunrise = sunrise;
     }
 
-    public Integer getSunset() {
+    public Double getSunset() {
         return sunset;
     }
 
-    public void setSunset(Integer sunset) {
+    public void setSunset(Double sunset) {
         this.sunset = sunset;
     }
